% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Responses for 103 Surveys}
\format{

}
\usage{
survey_data
}
\description{
A dataset containing each response that was made in 103 prediction markets which
were attempting to predict the outcome of direct replications. See; (Camerer et al., 2016, 2018; Dreber et al., 2015; Forsell et al., 2018)
}
\keyword{datasets}
